+++ main.go
@@ -2,16 +2,14 @@ package main
 
 import (
 	"context"
-	"crypto/sha256"
-	"crypto/sha512"
 	"fmt"
 	"github.com/IBM/sarama"
 	"github.com/getsentry/sentry-go"
-	"github.com/xdg-go/scram"
 	"log"
 	"net/http"
 	"os"
 	"os/signal"
+	"service_bus_consumer/common"
 	"service_bus_consumer/config"
 	"service_bus_consumer/handle"
 	"service_bus_consumer/kafka_target"
@@ -20,35 +18,6 @@ import (
 	"syscall"
 	"time"
 )
-
-var (
-	SHA256 scram.HashGeneratorFcn = sha256.New
-	SHA512 scram.HashGeneratorFcn = sha512.New
-)
-
-type XDGSCRAMClient struct {
-	*scram.Client
-	*scram.ClientConversation
-	scram.HashGeneratorFcn
-}
-
-func (x *XDGSCRAMClient) Begin(userName, password, authzID string) (err error) {
-	x.Client, err = x.HashGeneratorFcn.NewClient(userName, password, authzID)
-	if err != nil {
-		return err
-	}
-	x.ClientConversation = x.Client.NewConversation()
-	return nil
-}
-
-func (x *XDGSCRAMClient) Step(challenge string) (response string, err error) {
-	response, err = x.ClientConversation.Step(challenge)
-	return
-}
-
-func (x *XDGSCRAMClient) Done() bool {
-	return x.ClientConversation.Done()
-}
 
 type Consumer struct {
 	ready         chan bool
@@ -70,7 +39,7 @@ func (c *Consumer) ConsumeClaim(session sarama.ConsumerGroupSession, claim saram
 		select {
 		case message, ok := <-claim.Messages():
 			if !ok {
-				fmt.Println("message channel was closed")
+				logger.Error("message channel was closed")
 				return nil
 			} else {
 				handle.HandleData(message.Value)
@@ -102,7 +71,7 @@ func main() {
 		return
 	}
 	if err != nil {
-		fmt.Println("connect kafka subcriber: ", err)
+		logger.Error("connect kafka subscriber: ", err.Error())
 		return
 	}
 	conf := config.InitConfig()
@@ -119,10 +88,10 @@ func main() {
 		kafkaConfig.Net.SASL.User = conf.KafkaUserPublisher
 		kafkaConfig.Net.SASL.Password = conf.KafkaPasswordPublisher
 		if conf.KafkaMechanismPublisher == "SHA512" {
-			kafkaConfig.Net.SASL.SCRAMClientGeneratorFunc = func() sarama.SCRAMClient { return &XDGSCRAMClient{HashGeneratorFcn: SHA512} }
+			kafkaConfig.Net.SASL.SCRAMClientGeneratorFunc = func() sarama.SCRAMClient { return &common.XDGSCRAMClient{HashGeneratorFcn: common.SHA512} }
 			kafkaConfig.Net.SASL.Mechanism = sarama.SASLTypeSCRAMSHA512
 		} else if conf.KafkaMechanismPublisher == "SHA256" {
-			kafkaConfig.Net.SASL.SCRAMClientGeneratorFunc = func() sarama.SCRAMClient { return &XDGSCRAMClient{HashGeneratorFcn: SHA256} }
+			kafkaConfig.Net.SASL.SCRAMClientGeneratorFunc = func() sarama.SCRAMClient { return &common.XDGSCRAMClient{HashGeneratorFcn: common.SHA256} }
 			kafkaConfig.Net.SASL.Mechanism = sarama.SASLTypeSCRAMSHA256
 		} else {
 			sentry.CaptureException(fmt.Errorf("invalid SHA algorithm \"%s\": can be either \"sha256\" or \"sha512\"", conf.KafkaMechanismPublisher))
